Move QueryExecutor to query-core module
