+++ java/query/query-core/src/main/java/sleeper/query/core/recordretrieval/QueryExecutor.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.query.runner.recordretrieval;
+package sleeper.query.core.recordretrieval;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -31,8 +31,6 @@ import sleeper.core.util.ObjectFactory;
 import sleeper.query.core.model.LeafPartitionQuery;
 import sleeper.query.core.model.Query;
 import sleeper.query.core.model.QueryException;
-import sleeper.query.core.recordretrieval.LeafPartitionQueryExecutor;
-import sleeper.query.core.recordretrieval.LeafPartitionRecordRetriever;
 
 import java.time.Instant;
 import java.time.temporal.ChronoUnit;
