+++ java/query/query-core/src/test/java/sleeper/query/core/recordretrieval/QueryExecutorTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.query.runner.recordretrieval;
+package sleeper.query.core.recordretrieval;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
@@ -42,7 +42,6 @@ import sleeper.core.util.ObjectFactory;
 import sleeper.query.core.model.Query;
 import sleeper.query.core.model.QueryException;
 import sleeper.query.core.model.QueryProcessingConfig;
-import sleeper.query.core.recordretrieval.RecordRetrievalException;
 
 import java.io.IOException;
 import java.time.Instant;
