+++ java/query/query-runner/src/test/java/sleeper/query/runner/recordretrieval/QueryExecutorIT.java
@@ -51,6 +51,7 @@ import sleeper.query.core.model.LeafPartitionQuery;
 import sleeper.query.core.model.Query;
 import sleeper.query.core.model.QueryException;
 import sleeper.query.core.model.QueryProcessingConfig;
+import sleeper.query.core.recordretrieval.QueryExecutor;
 
 import java.io.IOException;
 import java.nio.file.Path;
